+++ hutool-cache/src/main/java/com/xiaoleilu/hutool/cache/impl/NoCache.java
@@ -1,6 +1,8 @@
-package com.xiaoleilu.hutool.cache;
+package com.xiaoleilu.hutool.cache.impl;
 
 import java.util.Iterator;
+
+import com.xiaoleilu.hutool.cache.Cache;
 
 /**
  * 无缓存实现，用于快速关闭缓存