+++ hutool-core/src/main/java/com/xiaoleilu/hutool/util/ClassUtil.java
@@ -25,7 +25,8 @@ import com.xiaoleilu.hutool.lang.Filter;
 import com.xiaoleilu.hutool.lang.Singleton;
 
 /**
- * 类工具类 1、扫描指定包下的所有类<br>
+ * 类工具类 <br>
+ * 1、扫描指定包下的所有类<br>
  * 参考 http://www.oschina.net/code/snippet_234657_22722
  * 
  * @author seaside_hi, xiaoleilu