+++ hutool-core/src/main/java/com/xiaoleilu/hutool/util/NumberUtil.java
@@ -169,7 +169,17 @@ public final class NumberUtil {
 		final BigDecimal b = new BigDecimal(numberStr);
 		return b.setScale(scale, roundingMode).doubleValue();
 	}
-
+	
+	/**
+	 * 保留小数位
+	 * @param number 被保留小数的数字
+	 * @param digit 保留的小数位数
+	 * @return 保留小数后的字符串
+	 */
+	public static String roundStr(double number, int digit) {
+		return String.format("%."+digit + 'f', number);
+	}
+	
 	/**
 	 * 格式化double
 	 *